feat: add colorize helper; refactor: extract handlers to modules
