+++ src/commands/handlers/attack.ts
@@ -1,0 +1,6 @@
+import { WSData } from '../interfaces/ws-data.interface';
+
+// TODO хендлить атаки не в свой ход
+export function handleAttack(_wsData: WSData, _data: string): void {
+  throw new Error('handleAttack not implemented');
+}
