+++ src/commands/handlers/random-attack.ts
@@ -1,0 +1,5 @@
+import { WSData } from '../interfaces/ws-data.interface';
+
+export function handleRandomAttack(_wsData: WSData, _data: string): void {
+  throw new Error('handleRandomAttack  not implemented');
+}
