+++ src/helpers/log.ts
@@ -1,11 +1,5 @@
 import { SocketMessage } from '../socket-message/interfaces/socket-message.interface';
-
-function cyan(str: string): string {
-  return `\x1b[36m${str}\x1b[0m`;
-}
-function green(str: string): string {
-  return `\x1b[32m${str}\x1b[0m`;
-}
+import { cyan, green, red } from './colorize';
 
 function logMessage(message: SocketMessage, isServerMessage: boolean): void {
   const [from, to] = (isServerMessage ? ['[Server]', 'client'] : ['[Client]', 'server']).map(cyan);
@@ -19,4 +13,8 @@ export function logRequest(message: SocketMessage): void {
 
 export function logResponse(message: SocketMessage): void {
   logMessage(message, true);
+}
+
+export function logError(err: unknown): void {
+  console.error(cyan('[Server]'), err instanceof Error ? red(err.message) : err);
 }
